==> s2_EI_linear_regression.ipynb <==
from sklearn.linear_model import LinearRegression
import numpy as np
import pandas as pd

#load training data
train_input_temp2 = pd.read_csv('EI_training_input.csv',header=None)
#train_output_temp = pd.read_csv('EI_training_output_data.csv',header=None)
train_output_temp2 = pd.read_csv('EI_training_output.csv',header=None)

train_application_temp = pd.read_csv('EI_application_input.csv',header=None)

train_input=np.array(train_input_temp2)
train_application=np.array(train_application_temp)
#train_input

train_output=np.array(train_output_temp2)#
train_output[[1,5],:]

model = LinearRegression().fit(train_input,train_output)

r_sq = model.score(train_input,train_output)
r_sq

y_pred = model.predict(train_input)
y_appl = model.predict(train_application)
y_pred

result      = y_pred.argmax(axis=1).astype(int)
result_appl = y_appl.argmax(axis=1).astype(int)
result

result_appl

output=train_output.argmax(axis=1).astype(int)

diff=result-output
diff

print(np.count_nonzero(diff),' over', len(diff),' samples')
