==> s3_EI_test_execute.ipynb <==
import sklearn
from sklearn.feature_extraction.text import CountVectorizer
import pandas as pd
import mxnet as mx
from mxnet import nd
from mxnet.gluon import nn
import numpy as np
import sys
import json

data = pd.read_csv('EI_validation_input.csv',header=None)
data

results = pd.read_csv('EI_validation_output.csv',header=None)
results

print("audit_deltaEI.data") 
p=pd.read_pickle("audit_deltaEI.data")

ctx = mx.cpu()
layer = nn.Sequential()
with layer.name_scope():
    layer.add(
        nn.Dense(6, activation="relu"),
        nn.Dropout(0.5),          # dropout to ease the overfitting problem
        nn.Dense(3)
    )
layer.load_parameters('audit_deltaEI.params',ctx)

predict=nd.array(data)
predict

test_result=layer(predict).argmax(axis=1).astype(int)
print('Prediction \n')
sys.stdout.write(json.dumps(test_result.asnumpy().tolist()))

audit_result=nd.array(results).argmax(axis=1).astype(int)
print('Audit results \n')
sys.stdout.write(json.dumps(audit_result.asnumpy().tolist()))

comp_result=test_result-audit_result
sys.stdout.write(json.dumps(comp_result.asnumpy().tolist()))
